feat(docs): create command-reference binary for documentation generation
