+++ packages/cli/package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "type": "module",
   "bin": {
-    "rover": "dist/index.js"
+    "rover": "dist/index.js",
+    "command-reference": "dist/command-reference.js"
   },
   "license": "Apache-2.0",
   "keywords": [
@@ -34,8 +35,8 @@
     "dist/*"
   ],
   "scripts": {
-    "build": "tsdown src/index.ts",
-    "dev": "TSUP_DEV=true tsdown --watch src/ src/index.ts",
+    "build": "tsdown",
+    "dev": "TSUP_DEV=true tsdown --watch src/",
     "start": "node dist/index.js",
     "clean": "rm -rf dist",
     "check": "tsc --noEmit",
